+++ test/module-types/CommonJS.test.cjs
@@ -1,0 +1,8 @@
+const assert = require('assert');
+const { describe, it } = require('../..');
+
+describe('CommonJS Suite', () => {
+  it('should run CommonJS Test Suites', async () => {
+    assert.ok(true);
+  });
+});