+++ test/module-types/ECMAScript.test.mjs
@@ -1,0 +1,9 @@
+import assert from 'assert';
+import zunit from '../../index.js';
+const { describe, it } = zunit;
+
+describe('ECMAScript Suite', () => {
+  it('should run ECMAScript Test Suites', async () => {
+    assert.ok(true);
+  });
+});